app/requirements_base.txt
# Core cross-platform dependencies for TheMusicBox
flask>=2.2.0
flask-cors>=3.0.10
python-dotenv>=1.0.0
python-socketio>=5.8.0
asgiref>=3.5.2
requests>=2.28.0
pytest>=7.0.0
gevent>=21.12.0
flask_socketio>=5.3.0
colorama>=0.4.4
mutagen>=1.45.0
# --- ASYNC MIGRATION ---
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
aiofiles>=0.8.0
yt_dlp>=2023.3.4
rx>=3.2.0
python-multipart>=0.0.5